# openapi.yaml
openapi: 3.0.0
info:
  title: Inventory Management API
  version: 1.0.0
  description: An API for managing inventory items.
paths:
  /items:
    get:
      summary: Retrieve a list of all items in the inventory.
      responses:
        '200':
          description: A list of items.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
    post:
      summary: Create a new item in the inventory.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
      responses:
        '201':
          description: The new item has been created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'

  /items/{id}:
    get:
      summary: Retrieve a specific item by its ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The item details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found.
    put:
      summary: Update an existing item.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Item'
      responses:
        '200':
          description: The item has been updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found.
    delete:
      summary: Delete an item from the inventory.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: The item has been successfully deleted.
        '404':
          description: Item not found.

components:
  schemas:
    Item:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier of the item.
        name:
          type: string
          description: The name of the item.
        description:
          type: string
          description: Description of the item.
        quantity:
          type: integer
          description: Current quantity of the item in stock.
        price:
          type: number
          format: float
          description: The price of the item.
        supplier:
          type: string
          description: The supplier of the item.